+++ src/notificationsWidget/notificationsWidgetWebview.ts
@@ -2,9 +2,13 @@ import { ExtensionContext } from "vscode";
 import { Capability } from "../capabilities/capabilities";
 import { fireEvent } from "../binary/requests/requests";
 import { StateType } from "../globals/consts";
-import registerWidgetWebviewProvider from "../widgetWebview";
+import registerWidgetWebviewProvider from "../widgetWebview/widgetWebview";
 
-function registerNotificaitonsWebviewProvider(context: ExtensionContext): void {
+const LOADED_NOTIFICATIONS_WIDGET = "loaded-notificaitons-widget-as-webview";
+
+export default function registerNotificationsWebviewProvider(
+  context: ExtensionContext
+): void {
   registerWidgetWebviewProvider(context, {
     capability: Capability.NOTIFICATIONS_WIDGET,
     getHubBaseUrlSource: StateType.NOTIFICATIONS_WIDGET_WEBVIEW,
@@ -13,10 +17,8 @@ function registerNotificaitonsWebviewProvider(context: ExtensionContext): void {
     readyCommand: "tabnine.notifications-ready",
     onWebviewLoaded: () => {
       void fireEvent({
-        name: "loaded-notificaitons-widget-as-webview",
+        name: LOADED_NOTIFICATIONS_WIDGET,
       }).catch(console.error);
     },
   });
 }
-
-export default registerNotificaitonsWebviewProvider;
