+++ src/tabnineTodayWidget/tabnineTodayWidgetWebview.ts
@@ -2,9 +2,13 @@ import { ExtensionContext } from "vscode";
 import { fireEvent } from "../binary/requests/requests";
 import { Capability } from "../capabilities/capabilities";
 import { StateType } from "../globals/consts";
-import registerWidgetWebviewProvider from "../widgetWebview";
+import registerWidgetWebviewProvider from "../widgetWebview/widgetWebview";
 
-function registerTabnineTodayWidgetWebview(context: ExtensionContext): void {
+const LOADED_TABNINE_TODAY_WIDGET = "loaded-tabnine-today-widget-as-webview";
+
+export default function registerTabnineTodayWidgetWebview(
+  context: ExtensionContext
+): void {
   registerWidgetWebviewProvider(context, {
     capability: Capability.TABNINE_TODAY_WIDGET,
     getHubBaseUrlSource: StateType.TABNINE_TODAY_WIDGET_WEBVIEW,
@@ -13,10 +17,8 @@ function registerTabnineTodayWidgetWebview(context: ExtensionContext): void {
     viewId: "tabnine-today",
     onWebviewLoaded: () => {
       void fireEvent({
-        name: "loaded-tabnine-today-widget-as-webview",
+        name: LOADED_TABNINE_TODAY_WIDGET,
       }).catch(console.error);
     },
   });
 }
-
-export default registerTabnineTodayWidgetWebview;
